# docs/OpenAPI3 - Struttura API.yaml
openapi: 3.0.1
info:
  title: AgroAPI
  version: 1.0.0
  description: API per i servizi agricoli della Regione Abruzzo
  contact:
    email: bryantsarabia@gmail.com
    name: Bryant Sarabia
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: "http://localhost:8000/api/v1"
    variables: {}
    description: Main server
...
paths:
...
  /reports:
    get:
      summary: >-
        Ritorna le segnalazioni dell'ultimo mese
      tags:
        - infestanti
      description: ''
      operationId: getReports
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pest'
              example:
                data:
                  - user_id: "21"
                    report_id: "1"
                    name: 'Hydrocotyle'
                    message: 'Vi consiglio di utilizzare X prodotto'
                    distance: 51.0121
                    coordinates:
                      lat: 43.1231212
                      lon: 10.1231242
                    created_at: '2020-10-10'
                  - user_id: "41"
                    report_id: "2"
                    name: 'Plantago'
                    message: 'Proteggete le vostre colture o saranno in pericolo'
                    distance: 51.0121
                    coordinates:
                      lat: 43.1231212
                      lon: 10.1231242
                    created_at: '2020-10-14'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '406':
          $ref: '#/components/responses/406NotAcceptable'
        '500':
          $ref: '#/components/responses/500Error'
security:
  - bearerAuth: []
  - ApiKey: []
components:
  schemas:
...
    Coordinate:
      type: object
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
      example:
        lat: '42.02531667'
        lon: '13.45065'
...
    Pest:
      type: object
      properties:
        user_id:
          type: string
          nullable: true
        report_id:
          type: string
        name:
          type: string
        message:
          type: string
        distance:
          type: number
          format: double
          nullable: true
        coordinates:
          $ref: '#/components/schemas/Coordinate'
        created_at:
          type: string
          format: date-time
          nullable: true
    Error:
      type: object
      properties:
        code:
          type: integer
        title:
          type: string
        details:
          type: string
...
  responses:
    500Error:
      description: Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            title: Server error
            details: ''
...
    400BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            title: Bad request
            details: ''
    401Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            title: Unauthorized
            details: 'Incorrect API Key, please try again'
    406NotAcceptable:
      description: Missing Accept header
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 406
            title: Not acceptable
            details: Be sure to set the header "Accept" to "application/json"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key